Updated the enums types from int to uint8_t

+++ JsonMidiCreator/TalkieEthernetENC_SPI/src/JsonMessage.hpp
@@ -30,7 +30,7 @@ using MessageValue = TalkieCodes::MessageValue;
 class JsonMessage {
 public:
 	
-	enum ValueType : int {
+	enum ValueType : uint8_t {
 		STRING, INTEGER, OTHER, VOID
 	};
